Create LoginDto Add LoginAsync Method to IAuthManager Interface Implement LoginAsync in AuthManager Class Create Login Endpoint in UsersController
