+++ HotelListing.API/Contracts/IAuthManager.cs
@@ -6,5 +6,6 @@ namespace HotelListing.API.Contracts
     public interface IAuthManager
     {
         Task<IEnumerable<IdentityError>> RegisterAsync(ApiUserDto apiUserDto);
+        Task<bool> LoginAsync(LoginDto loginDto);
     }
 }
