+++ HotelListing.API/Models/ModelsDto/ApiUser/ApiUserDto.cs
@@ -3,23 +3,16 @@ using System.ComponentModel.DataAnnotations;
 
 namespace HotelListing.API.Models.ModelsDto.ApiUser
 {
-    public class ApiUserDto
+    public class ApiUserDto : LoginDto
     {
         [Required]
         public string UserName { get; set; }
-
-        [Required]
-        [EmailAddress]
-        public string Email { get; set; }
 
         [Required]
         public string FirstName { get; set; }
 
         [Required]
         public string LastName { get; set; }
-
-        [Required]
-        public string Password { get; set; }
 
         [Required]
         [Phone]
